requirements.txt
crewai
streamlit
crewai-tools
python-dotenv

